[#1889] fix: Call destructor of BufferControlBlock in every case
If created via placement new, the memory that contains the control block
is managed separately from the controlblock. Therefore delete does not
have to be called.

However, in this case, we have to call the BufferControlBlock's
destructor manually before releasing the memory.

Even though currently the BufferControlBlock's destructor is defaulted,
I add the previously missing call to it's destructor in case things
change in the future.