[#2318] docker entrypoint: add option to include perf for current kernel
CLion's profiling uses perf inside the docker container when using
remote development.
Perf needs to be compiled for the current kernel version to work.
Since the kernel is defined by the host system, its version can only be
determined at runtime, thus relying on the entrypoint script here.

To enable the container needs to be started with:
-e PROFILER_ENABLED=true --cap-add SYS_PTRACE --cap-add SYS_ADMIN