[#3998] Changed how SliceMergeTask object creation is done within tuple buffers.
Previously, the SliceMergeTask object was being directly created from the buffer. However, this causes issues when the buffer is not sufficient or not correctly allocated. To happen these issues, the change initializes the SliceMergeTask object using the “new” operator. This ensures that even if the buffer does not have sufficient storage, a new storage area will be allocated separately to initialize the SliceMergeTask object, preventing a possible crash or corruption of data.

Signed-off-by: Philipp Grulich <philippgrulich@hotmail.de>