#4180 Common Interface for Inference
InferenceAdapter is a common base class for TensorFlowAdapter and ONNXAdapter.
It offers methods to insert inputs, retrieve outputs and invoke the model inference. Dispatching to
specific runtimes is handled withing the `InferModelHandler`

The base class internally offers a `std::span` to subclasses when inference is invoked.

This commit also removes/changes previous C-Code inside the TensorflowAdapter to C++.