Unikernel Export: Proof of Concept
Unikernel Export Application accepts a YAML input file, which specifies
a topology and a query. The Exporter will run the parsing,
optimization, placement, pipelining, and compilation phases until it
reaches the MLIR Backend. The Exporter exports the MLIR Operator into
an object file that exposes the UnikernelStage ABI, which is statically
linked against by the Unikernel. Additionally, Information regarding
the placement (IP, ports, partitions) is exported as a YAML file as an
input for the unikernel generator and the Unikernel Source and Sink
applications.
The Exporter tries to avoid using unnecessary components of NES where
possible (i.e., no runtime components). NodeEngine, QueryManager are
replaced with NoOp implementations.
OperatorHandlers are replaced with stubbed implementations, which are
useful to keep track of which of the handlers were instantiated with
which parameters.

- Compare Memory Addresses of returned OperatorHandlerPtr to identify Handlers
 that are shared across pipelines. E.g. Both Join Build Sides and the actual join
 will use the same JoinOperatorHandler, which contains the state of the join
- Shared operatorhandlers are instantiated in a separate compilation unit sharedHandlers.o
- the Stage::getOperatorHandler will test if the index is one of the shared operator handler and call getSharedOperatorHandler