Review: Recursive Locking in StreamJoinOperatorHandler
The Unikernel does not submit the Buffer into a QueryManager, instead it directly
calls the next Pipeline. For Joining this introduces a problem, where emit is
called while still holding the Exclusive Lock. This effectively, deadlocks the
Thread, because the Probe Phase also requires the Lock, which it already holds.

I circumvented the issue by copying/moving all slices that are to be emitted,
after marking them as `EMITTED_TO_PROBE`. This allows the building Stage to
release the lock before calling the probing Stage.