[#4323] Topology graph queries should return collection of Nodes.
A Topology in NES is not a Tree, queries such as LCA can potentially
return multiple nodes. This is now reflected in the API. There has
never been any specification of the order of Nodes returned by
`findNodesInBetween`, an additional sort function allows the set
to be turned into a sorted vector.