4656 add statistic id into system (#4663)
## What is the purpose of the change

This PR adds a `StatisticId` into our system. Any component, e.g.,
physical source or an operator, can acquire a `StatisticId`. With this
`StatisticId` and a metric, i.e., cardinality over a field, we can then
uniquely identify what statistic is built for what.
Currently, any logical operator and every physical source receives a
`StatisticId` during the `StatisticInferencePhase` and the
`LogicalSourceExpansionRule`. This statistic id is then lowered down to
the physical and then to the nautilus operator. Similar to the
`OriginId`, the `StatisticId` is moved around in our system as part of
the tuple buffer metadata.


## Brief change log

- Changes for the `StatisticId` in the operator
  - Added `StatisticId` in the logical `Operator` 
- Added `StatisticId` in the `PhysicalOperator` and touched all child
classes of `PhysicalOperator` to accomodate the new `StatisticId`
- Added `StatisticId` in the `DataSource` and touched all sources to
accomodate the new `StatisticId`
  - Added new `StatisticInferencePhase` to the optimizer
- Adapted existing `LogicalSourceExpansionRule` to read the
`StatisticId` from the source catalog

- Changes for the `StatisticId` in the tuple buffer
- Added new getter and setter for the `StatisticId` in the
`BufferControlBlock`
- Added new getter and setter for the `StatisticId` in the `TupleBuffer`
- Added new getter and setter for the `StatisticId` in the
`ExecutionContext`
  
- Changes for the `StatisticId` in the query execution part
- Each operator writes its `StatisticId` in the `ExecutionContext`, with
the exception of the `Scan`. It writes the `StatisticId` of the tuple
buffer.
- The `Emit` operator writes then the `StatisticId` to the tuple buffer
  
  
- Changes to the query API and the statistic build operator
  - MetricHash is now part of the statistic build operator
  - Adapted `StatisticKey` to consist of the metric and the statistic id


## Verifying this change

This change added tests.

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
  - The compiler: no
  - The threading model: no
  - The Runtime per-record code paths (performance sensitive): no
  - The network stack: no
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): no

## Documentation

  - Does this pull request introduce a new feature? no

## Issue Closed by this pull request:

This PR closes #4656, closes #4655, and closes #4657

---------

Co-authored-by: Nils Schubert <n.schubert.1@tu-berlin.de>