StatisticCoordinator track and probe implementation (#4683)
## What is the purpose of the change

This PR implements the `trackStatistic` and `probeStatistic` of the
`StatisticCoordinator`. We return now a vector of `StatisticKeys` that
can then be used for probing the statistic.
We require a vector as one track request can result in multiple
`StatisticIds`, for example due to the logical source expansion rule.
Furthermore, we have created a new class `ProbeExpression` that is used
to evaluate a statistic and retrieve a `StatisticValue`. One example
might be that we create a sample (`Statistic`) and now want to know the
selectivity of `f1 < 10`.


## Verifying this change

This change added tests.

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
  - The compiler: no
  - The threading model: no
  - The Runtime per-record code paths (performance sensitive): no
  - The network stack: no
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): no

## Documentation

  - Does this pull request introduce a new feature? no
  
## Issue Closed by this pull request:

This PR closes #4654, closes #4653, and closes #4633.

---------

Co-authored-by: Nils Schubert <n.schubert.1@tu-berlin.de>