[#4352] added error for time data type (#4400)
## What is the purpose of the change

The datatype of the time used for the window operator in a query has to
be an Integer. Previously the error message did not specify this. Now
the necessary changes are clarified in the error.


## Brief change log
 
The type of time is now checked and if it's not int a detailed error is
printed and an exception is thrown

## Verifying this change

This change is a trivial rework and can be verified with the
`E2ECoordinatorSingleWorkerTest.cpp`.

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
  - The compiler: no
  - The threading model: no
  - The Runtime per-record code paths (performance sensitive): no
  - The network stack: no
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): no

## Documentation

  - Does this pull request introduce a new feature? no

## Issue Closed by this pull request:

This PR closes #4352

---------

Co-authored-by: Nils Schubert <n.schubert.1@tu-berlin.de>