[#4842] actually fix include guards (#4853)
## What is the purpose of the change

This is the sibling PR to #4845. A separate PR is created two have two
separate commits: #4845 containing all changes to the build target, and
this PR/commit with the result of the run.

## Brief change log

Fixed include guards by running `make fix-guards`

## Verifying this change

This change is a trivial rework / code cleanup without any test
coverage.

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
  - The compiler: no
  - The threading model: no
  - The Runtime per-record code paths (performance sensitive): no
  - The network stack: no
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): no

## Documentation

  - Does this pull request introduce a new feature? no

## Issue Closed by this pull request:

This PR relates to #4842.