4723 helpful failure message no workers (#4892)
## What is the purpose of the change

This change was created to give stricter requirements for the
numWorkerThreads parameter of configurations.
If numWorkerThreads is 0, validation should fail, and an error message
should be printed.

In the issue, the request was also made to fail the worker if this error
is encountered.
I'm unsure how to test this so far, and would like to know if the
exception thrown by the validator already meets this requirement.

### Follow Up Work / Discussion
I'd have liked to provide a more useful error message depending on which
validator has failed, but this would require some extra out-of-scope
changes. Other parameters also use the existing validator which allows
for 0s, and should probably also be moved.

I've added a new validator to replace an existing one, but would have
preferred to change the validator logic to allow chaining of multiple
validators for a field. This would help reduce the number of validators
we create in total, and choose more expressive names for the validators.


## Brief change log

- added new validator to only allow positive integers
- set numWorkerThreads to use this validator instead of the existing one
which also allowed 0


## Verifying this change

This change added tests and can be verified as follows:
 - run the nes-configuration unit-tests
- a new change in the unit test for integer values confirms that setting
numWorkerThreads to 0 fails parameter validation

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): (yes / **no**)
  - The compiler: (yes / **no** / don't know)
  - The threading model: (yes / **no** / don't know)
- The Runtime per-record code paths (performance sensitive): (yes /
**no** / don't know)
  - The network stack: (yes / **no** / don't know)
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): (**yes** / no / don't
know)

## Documentation

  - Does this pull request introduce a new feature? (**yes** / no)
- If yes, how is the feature documented? (not applicable / docs /
JavaDocs / **not documented**)


## Issue Closed by this pull request:

closes #4723