4806 Fix error handling in the FileSink (#4903)
## What is the purpose of the change

This PR fixes error handling in the FileSink so that it does not crash
the worker. Fixes #4806.

## Brief change log

- Move setup code of the FileSink from the constructor to the `setup`
function and remove calls to `NES_ASSERT`
- Changed how `ExecutionResult::Error` is handled by
`ThreadPool::runningRoutine` and updated
`DynamicQueryManager::processNextTask` to notify the coordinator of a
task failure (see below)
- Updated `WorkerContext::waitForAsyncConnection` to handle case that a
reverse event channel future does not exist (see below)
- Updated documentation and log messages in `FileSink.cpp`
- Updated and refactored `TestHarness.cpp` (see below)

## Changes to `ThreadPool::runningRoutine` and
`DynamicQueryManager::processNextTask`

Previously, when `DynamicQueryManager::processNextTask` would return
`ExecutionResult::Error`, `ThreadPool::runningRoutine` would exit the
task execution loop and the node engine would stop.

This PR changes `ThreadPool::runningRoutine` so that no action is taken
for `ExecutionResult::Error`. Instead,
`DynamicQueryManager::processNextTask` calls `notifyTaskFailure` which
causes the query to be undeployed.

## Updates to `WorkerContext::waitForAsyncConnection`

The previous code assumed that there is always a reverse event channel
future for an operator ID. However, because we notify the task failure
it can happen that this code is executed twice. Once for the failure
event and once for the soft stop because the source has finished. See
the following log output:

```
[18:55:09.127353] [D] [thread 4347519] [NetworkSource.cpp:162] [reconfigure] NetworkSource: reconfigure() called 3::36::0::0 for the type FailEndOfStream
...
[18:55:04.087734] [D] [thread 4347347] [NetworkSource.cpp:162] [reconfigure] NetworkSource: reconfigure() called 2::21::0::0 for the type SoftEndOfStream
[18:55:04.087737] [W] [thread 4347347] [WorkerContext.cpp:232] [waitForAsyncConnectionEventChannel] Did not find a reverse event channel future; operatorId = 21
[18:55:04.087738] [W] [thread 4347347] [NetworkSource.cpp:251] [reconfigure] Did not find a reverse event channel future; operatorId = 21
```

When `reconfigure` is called for the second time, the reverse event
channel future has been removed. We log a warning messages in this case.

## Updates to `TestHarness.cpp`

- Added builder pattern methods to set the path of the test file sink
and set the append mode.
- Extracted code to (1) add the file sink to the query, (2) enqueue the
query, and (3) to stop the coordinator and workers, so that they can be
called individually from tests.
- Added a delegator method to `TestUtils::checkFailedOrTimeout`.

## Verifying this change

- Added integration tests in `FileSinkIntegrationTest.cpp`

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
  - The compiler: no
  - The threading model: unsure
  - The Runtime per-record code paths (performance sensitive): no
  - The network stack: no
  - Anything that affects deployment or recovery: yes

## Documentation

  - Does this pull request introduce a new feature? no
  - If yes, how is the feature documented? not applicable