4935 Fix serialization of data types in json (#4937)
## What is the purpose of the change

This PR fixes the serialization of data types other than strings in JSON
output.

Previously, all fields were serialized as strings. This prevents
interoperability with other tools that expect proper numbers.

Fixes #4935.

## Brief change log

* Updated `Iterator::dataJson()` to write numbers without quotes and
renamed it to `Iterator::serializeTupleAsJson()`
* Removed JSON exception handling in `Iterator::dataJson()`. Previously,
a JSON exception with logged as an error but then an empty string was
returned. The exception should now bubble up to query execution and
trigger a failure of the query.

## Verifying this change

* New test `IteratorTest::testJsonDataTypes`

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
  - The compiler: no
  - The threading model: no
  - The Runtime per-record code paths (performance sensitive): no
  - The network stack: no
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): (yes / no / don't know)

## Documentation

  - Does this pull request introduce a new feature? no
  - If yes, how is the feature documented? n/a