4967 task testing for times operator (#4979)
## What is the purpose of the change

This pull request tests if the TIMES-operator is translated properly to
the expected query-plan by comparing the structure of all different
kinds of times-queries with the expected query-plans.


## Brief change log
Tests for:
- TIMES without specified min/max occurrences
- TIMES with specified exact number of occurrences
- TIMES with 0 occurrences
- TIMES with specified min and max occurrences


## Verifying this change

This PR introduces no change to NES but adds additional test cases to
check the correctness of the AND-operator.

## Does this pull request potentially affect one of the following parts:

- Dependencies (does it add or upgrade a dependency): (no)
- The compiler: (no)
- The threading model: (no)
- The Runtime per-record code paths (performance sensitive): (no)
- The network stack: (no)
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): (no)

## Documentation

- Does this pull request introduce a new feature? (no)

## Issue Closed by this pull request:

This PR closes #4967