feat(Build) Prevent building shared libraries with undefined symbols
Previously mold did not refuse linking a shared library if not all symbols were
resolved, which is not uncommon for a linker to do. Currently we don't use any
runtime discovery of symbols, thus all symbols should be present during build
time. This shifts the error message into the build time instead of a dynamic
loader error during runtime.