refactor(Memory) Use the new Memory modules interfaces
Previously most code depended on a concrete BufferManager, this made the code
inflexible and hard to test. Unless the code explicitly need both a
PoolProvider and a BufferProvider this commit replaces the concrete
BufferManager class, with either of the interface.

Additionally many shared pointers have been replaced with plain references.