chore: replace vendored mini-yaml with yaml-cpp
mini-yaml was vendored and is stale since 2019.

In difference to mini-yaml, yaml-cpp throws when accessing
a non-existing key or when accessing as a bad type,
thus access checks were rewritten.

Also, yaml-cpp's iterator semantics differ a little,
sequence iterators are not tuples and the map iterator's
'.first' is a YAML::node, not a string.