feat(Development) Adds local environment setup script
The script creates the nebulastream/nes-development:local image
which copies the current user into the docker image preventing
any sorts of conflicts.

The script calculates the dependency-hash and by default attempts
to fetch a pre-built dependency image from the nebulastream docker
registry. The '-l' flag enables building the complete set of docker
images locally, if they are not yet provided in the registry.