feat(ErrorHandling) Use lazy stack traces
Currently, we generate the full exception stacktrace whenever
we create an exception. This is incredible expensive, especially
with static linking. The benchmark shows that generating a
stacktrace with resolved symbols takes seconds!

This commit enables lazily resolving the stack trace only if
it is required. To accomplish this the exception class inherits
from `cpptrace::lazy_exception` which constructs a `raw_stacktrace`
during construction, which is significantly cheaper.