#5082: Runtime error when WITHIN is expected but missing (#5087)
## What is the purpose of the change

In the PSL, the WITHIN keyword is optional in the grammar. There are a
few queries in which the WITHIN keyword in not needed so we can't just
make it mandatory. Instead, this PR added an error if WITHIN is expected
(eg if there is an AND or SEQ operator in the pattern).


## Brief change log

  - if pattern has an operator, throw an error
  - test for operators (for SEQ, AND, TIMES)


## Verifying this change

This change is a trivial work which changes can be verified by the test.

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
  - The compiler: no
  - The threading model: no
  - The Runtime per-record code paths (performance sensitive): no
  - The network stack: no
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): no

## Documentation

  - Does this pull request introduce a new feature? no

## Issue Closed by this pull request:

This PR closes #5082