fix(Nebuli) Fixed failing parsing of queries
I expect due to some caching behavior a bug occurred that we do not add all necessary header files to the file set for `nebuli`.

This fix changes the macro `add_nebuli_header` to also allow files to be copied to the file set. Aditionally, it adds the missing files to the file set via invoking the macro `add_nebuli_header`.