5052 slices shared join (#5099)
## What is the purpose of the change

Joins could be shared in NebulaStream if they have the same syntax
(window definition and join condition). However they could be deployed
at a different time which will need a new strategy to calculate slice
edges.
Example: Join A with range 10seconds and slide of 5seconds deployed at
t=0s; Join B the same but deployed at time=3s
The slices would need to change from {[0-5],[5-10],[10-15]..} to
{[0-3],[3-5],[5-8],[8-10],[10-13]..}

## Brief change log
- sliceAssigner.cpp: supports multiple deployment times of this join to
calculate slice edges for incoming tuples.
- StreamJoinOperatorHandlerSlicing.cpp: supports assigning slices to
windows of all deployment times
   - StreamJoinOperatorHandler.cpp:
- 1. can be reused by another query to gain access to the slices that
were created by this query and not yet triggered. This helps emulates a
situation in which we would not need to redeploy a new query, but could
adjust the running query.
- 2. adjusts the edges of already created slices once a new query
deployment time is added (does not change tuples inside of slice, so
tuples are potentially in the wrong slice after this PR


## Verifying this change

This change added tests and can be verified as follows:

- Added tests similar to streamJoinExecutionTest, but for a shared join
scenario
- Tests cover sliding and tumbling windows, as well as different
scenarios that could exist once a new query would be deployed.

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
- The compiler: don't know (does not change nautilus; the *compilation*
that is affected is from LQP to EQP)
  - The threading model: no
- The Runtime per-record code paths (performance sensitive): no
(performance relevant but for a case that was not supported before
[sharing joins])
  - The network stack: no
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): no

## Documentation

  - Does this pull request introduce a new feature? yes
  - If yes, how is the feature documented? JavaDocs

## Issue Closed by this pull request: 
This PR closes #5052