refactor(ParserCSV): improved stringField to formatted field conversion.
Prior, for every string field that we parsed, we traversed a big switch statement to check what type the specific field has and to then call the correct conversion function and cast correctly. Now, we scan the schema once, when we create the parser, and fill a vector with the correct conversion function for each field. During parsing, we can access that vector, get the correct conversion function and convert the current string field to the correct internal data type.