feat(Parser): parser to nes-source-parsers;no more TestTupleBuffer
Prior, the (source) parser was part of nes-sources. Now, there is a new component 'nes-source-parsers' that holds all source parser implementations. Right now, nes-sources depends on nes-source-parsers and source implementations still use the parser. In subsequent commits, we will move the parser implementation to the source thread. In subsequent issues, we will move the parser out of nes-sources completely, getting rid of the dependency.
With this commit, the parser does not use the TestTupleBuffer anymore, but writes directly to a TupleBuffer.