chore(Parser): Error code of from_chars now triggers exception.
The [std::from_chars](https://en.cppreference.com/w/cpp/utility/from_chars) function returns an error code if it was unable to parse an integer, e.g., if there was a leading whitespace (" 123") or an unrecognized symbol ("1-23"). We now handle this error code by throwing an exception as described in #500.