feat(SLT) Enhance Result Checker to support relative errors
Added a sink schema that gets defined similar to the sources. The defined sinks can then be used in queries. The defined schema of the sink is how the result tuples should be written down. This allows us to compare the result not via a simple string comparison but via their actual data type.