#5114 slice tuples shared join (#5120)
## What is the purpose of the change

A syntactically equal join produces joins over different time frames if
the queries are deployed at different time. We already adjusted slice
start and stop times to work in such an environment in Task #5052.
However since we deal with out-of-order tuples and won't know that a new
query arrives, some tuples will be in a wrong slice once a new query is
deployed. So we need to change the tuples inside slices to work with the
old model or check tuples of slices again in the probe phase.

To achieve this three approaches are proposed and this PR implements the
first approach:

Approach: Leave all slices that already contain tuples unchanged. Ensure
that these slices get probed for each window that potentially has a
tuple in the slice. If one of these slices gets probed, skip tuples that
are outside of the window definition.

## Brief change log

- Changed slice id to be a uint64_t that gets incremented and then
assigned to a new slice in the streamJoinOperatorHandler
- Added a timeFunctionPtr to probe operators
- Added functionality in nljProbe to check if tuples should be probed
(look at purpose)
- Added functionality to more closely resemble removing of a quey from a
shared streamJoinOperatorHandler
- Added access to the functionality of adding and removing queries to a
shared streamJoinOperatorHandler from an executablePipeline


## Verifying this change

tests in StreamJoinQuerySharedExecutionTest

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
  - The compiler: no
  - The threading model: no
- The Runtime per-record code paths (performance sensitive): no (it does
change for probe, but more or less only in the case of a shared join and
that was not supported before)
  - The network stack: no
  - Anything that affects deployment or recovery: no

## Documentation

  - Does this pull request introduce a new feature? (yes )
  - If yes, how is the feature documented? JavaDocs
  
## Issue Closed by this pull request:

This PR closes #5114