#5047 propagate markers along data stream (#5112)
## What is the purpose of the change
This PR introduces a reconfiguration marker that is propagated through
the data stream. This reduces the amount of messages between worker and
coordinator that are required to reconfigure a query (e.g. for operator
migration).

## Brief change log
The reconfiguration marker carries a set of reconfiguration events, that
define actions to be taken for certain decomposed query plans. On
receiving a marker, a network source will check if any contained event
is aimed the decomposed query plan which the source is part of. If so,
it will take the required actions and propagate the marker. Otherwise it
will ignore the marker and not propagate it.

- Use decomposed query id instead of id-string to match events to
decomposed plans
- implement propagation of reconfiguration marker from sources, through
executable pipelines to sinks
- Allow EndOfStream messages to carry additional metadata necessary for
marker propagation over the network
- Extend zmq server and exchange protocol to handle markers
- Implement handling of Drain-events at sources, executable pipelines
and sinks


## Verifying this change

This change added tests and can be verified as follows:

- added unit test that tests propagation of markers and shutting down a
query via a reconfiguration marker

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
  - The compiler: no
  - The threading model: no
  - The Runtime per-record code paths (performance sensitive): no
  - The network stack: yes
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): yes

## Documentation

  - Does this pull request introduce a new feature? yes
  - If yes, how is the feature documented? not documented

## Issue Closed by this pull request:

This PR closes #5047

---------

Co-authored-by: flang <flang@asian-lion.dima.tu-berlin.de>