Interval join tag 1 (#5041)
## What is the purpose of the change

This pull introduces the interval join operator to NES. 
The Interval Join joins two events on a common key where the timestamp
of an event b1 from stream B lies between a timestamp interval of event
a1 from stream A --> b.timestamp ∈ [a.timestamp + lowerBound;
a.timestamp + upperBound] or a.timestamp + lowerBound <= b.timestamp <=
a.timestamp + upperBound

## Brief change log
- *We have to specify the query for requesting an interval join; we,
therefore, updated the query.cpp to be able to handle the specific
keywords and the QueryPlanBuilder.*
- *To serialize the Interval Join Operator to be exchanged between
computing nodes, it has to be (de-)serialized. To this end, the .proto
files has been adapted. Serialization is part of the grpc module &
nes-operators -> Operators ->Serialization -> OperatorSerializationUtil*

  - *Physical Operators are created*
  - *Nautilus Operators are created*


## Verifying this change

This change added tests and can be verified as follows:

- SerializationUtilTest.cpp

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): (no)
  - The compiler: (yes )
  - The threading model: (no)
  - The Runtime per-record code paths (performance sensitive): (no)
  - The network stack: (no)
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): (no)

## Documentation

  - Does this pull request introduce a new feature? (yes)
- If yes, how is the feature documented? // link to project report, add
to nes doc

## Issue Closed by this pull request:

This PR closes #4824

---------

Co-authored-by: Benji33 <benjiblank95@gmail.com>
Co-authored-by: Maximsch01 <73937394+Maximsch01@users.noreply.github.com>
Co-authored-by: esracosgun <esra.cosgun@gmx.de>