feat(OptionalPlugins): defining signature of registry in one place.
We now define the return type and the arguments for a registry interface in the header that instantiates the registry, e.g, SourceRegistry.hpp. Everywhere else, we access the definitions from the Registry header, including the generated registrars. As a result, the registrar is zero-config, making it easier to use and more straight forward to generate.