feat(Engine) Adds PendingPipelineStop
This commit adds a reference counter to each pipeline, which would prevent
a pipeline stop from stopping the pipeline while there are still pending tasks.

Using a separate reference counter allows a failure to ignore the pending ref
counter and proceed with stopping the pipeline. Future tasks are skipped as the
pipeline has expired.