#5115 slices tuples shared join approach two (#5142)
## What is the purpose of the change

A syntactically equal join produces joins over different time frames if
the queries are deployed at different time. We already adjusted slice
start and stop times to work in such an environment in Task #5052.
However since we deal with out-of-order tuples and won't know that a new
query arrives, some tuples will be in a wrong slice once a new query is
deployed. So we need to change the tuples inside slices to work with the
old model or check tuples of slices again in the probe phase.

To achieve this three approaches are proposed and this PR implements the
second approach:

Approach: Check all slices that potentially contain tuples after the
deployment of the new query. For each slice adjust the start and end
time to what it would have been if the other query was already deployed
at its deployment time. Now go through each tuple from the slice and if
the tuples ts is outside of the slices start and end time, remove this
tuple and put it in its correct slice. Afterwards remove empty spaces
that might have appeared in slices, by moving all tuples that stayed
inside the slice adjacent to each other.


## Brief change log
- added functionality that delete wrong records from slices (complete
record or keeping dangling text part until complete slice is removed)
- added time function to streamJoinOperatorHandler
- mini toString improvement in record

## Verifying this change
- adjusting existing tests in StreamJoinQuerySharedExecutionTest.cpp
- new tests (especially for text part of records) in
PagedVectorVarSizedTest.cpp

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): no
  - The compiler: (no
  - The threading model: no 
  - The Runtime per-record code paths (performance sensitive): no
  - The network stack: no
- Anything that affects deployment or recovery: Coordinator (and its
components), NodeEngine (and its components): no

## Documentation

  - Does this pull request introduce a new feature? yes
  - If yes, how is the feature documented? JavaDocs

## Issue Closed by this pull request:

This PR closes #5115