chore(Execution) Executable Functions:
Executable Functions are now purely reading from a record and never writing/updating the provided record. This is similar to a mathematical function that receives an input (in our case a record) and then returns a new value (in our case a VarVal). Additionally, this allows us to provide the record as a const reference, allowing us to write more const in our code.