feat(Execution) BufferProvider for Pipeline:
We have changed the way how an operator has access to a buffer provider. Before, we have stored one buffer provider in the operator handler for the whole duration of the query. Now, we store it in the Execution Context. Giving us the chance of swapping the buffer provider to another one, if we please to do so.