feat(TestUtils) Nautilus Test Utils:
We have added a new nautilus test utils that should help us in writing tests for nautilus code, e.g., test for a nautilus wrapper. Additionally, we have used this new test utils to have the chained hash map and paged vector be tested in the interpreter as well as the compiled approach.