#574 : Replaced the DataTypeFactory with a DataTypeRegistry
This is part of the nes-data-types refactoring, which includes changing our external numeric representation to types with preset bounds and bits.

Due to this change, the bug where we were not able to capture the upper bound of a UInt64 correctly has been fixed, since we can store upper bounds of integers as uint64_t now.
Previously, Integers were able to have a negative upper bound. This is no longer the case, as we only represent logical types of 8, 16, 32 and 64 bit integers signed and unsigned.

This included the removal of all data type factory functions which tightened the lower and upper bounds of integer/float data types. These functions were currently unused in the code, aswell as the usage of the creation of floats with custom bounds. Therefore, only the creation of 32-bit floats and 64-bit doubles is supported now.

This includes the removal of any function which allows integer data types to have modified bits or upper/lower bounds beyond the 8, 16, 32, 64 signed and unsigned bit default values.
This change is also reflected in the deserialization process of the integer data type.