feat(SequenceShredder): handling implicit TupDel of first buffer
Prior, the four functions that find sequence numbers (upper/lower + no-wrapping/wrapping) set the sequence number to '0' if it was invalid. Now, they return a boolean flag, indicating whether the sequence number is valid or not.
That allows us to handle the dummy '0' sequence number as a valid sequence number.