fix(CI): code-cov: fix pr caching
Previously, the "Copy code coverage of changed files" step was run even
on cache hit, but the preceding step it depends upon.
This cause the workflow to fail when hitting cache, e.g. when re-running
it.

This fix extends the caching from the .lcov file to the whole folder,
that also contains the generated html, so now it is also cached and the
"Copy code.." step (and its dependency step) can also be skipped.

For clarity, the folder is renamed from `/tmp/cache` to `/tmp/cov-cache`