feat(InputFormatterTask): Introduces new Interface for InputFormatters.
Prior, the CSVInputFormatter contained essentially all of the logic to parse raw input buffers.
This is bad, because every new formatter needs to reimplement a lot of the logic to format input buffers.
With this commit, most of the logic moves into the InputFormatterTask.
InputFormatters now only need to implement two functions.
One, which takes a buffer and writes all field offsets to a special data structure. A second, which parses a single spanning tuple and writes the field offsets to a pointer.