fix(CI): fix check_todos.py "no merge base found"
Previously, the script would sometimes break with

    fatal: no merge base found

maybe when not based directly on main.

It could be that since `BASE_REF` would point to main but the checkout
action would only fetch commits from HEAD to the merge base, thus the
sha of main would be unknown, causing a failure.

Anyhow, this should prevent the failure.