chore(SyncInputFormatterTask): Introduced SyncInputFormatterTask.
The SyncInputFormatterTask gets buffers in order from the source. Consequently, formatting becomes significantly simpler, since all prior data is available, including all metadata produced by prior threads. The SyncInputFormatterTask has a Notifier that it can use to notify the source to put the next data buffer into the TaskQueue. Until the SyncInputFormatterTask notifies the Source, it is guaranteed to be the only thread that has access to the state of the InputFormatter and its own state.